Rename all DAO implementation classes, remove prefix MySQL
