+++ src/main/java/by/training/payment/dao/impl/ParameterHeaderDAOImpl.java
@@ -6,60 +6,59 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-import by.training.payment.dao.RequestStatusDAO;
-import by.training.payment.entity.RequestStatus;
+import by.training.payment.dao.ParameterHeaderDAO;
+import by.training.payment.entity.ParameterHeader;
 import by.training.payment.exception.DAOException;
 import by.training.payment.pool.PoolConnection;
 import by.training.payment.pool.ProxyConnection;
 
-public class MySQLRequestStatusDAO extends SQLUtil implements RequestStatusDAO {
+public class ParameterHeaderDAOImpl extends SQLUtil implements ParameterHeaderDAO {
 
-	private static final String GET_REQUEST_STATUS_BY_VALUE = "SELECT * FROM request_status WHERE request_status_value = ?";
-	private static final String GET_REQUEST_STATUS_LIST = "SELECT * FROM request_status";
+	private static final String GET_PARAMETER_HEADER_BY_NAME = "SELECT * FROM parameter_header WHERE parameter_header_name = ?";
+	private static final String GET_ALL_PARAMETER_HEADERS = "SELECT * FROM parameter_header";
 
 	@Override
-	public List<RequestStatus> getRequestStatusList() throws DAOException {
-		List<RequestStatus> requestStatus = new ArrayList<>();
+	public List<ParameterHeader> getAllParameterHeaders() throws DAOException {
+		List<ParameterHeader> parameterHeaders = new ArrayList<>();
 		PreparedStatement statement = null;
 		ResultSet resultSet = null;
 		try (ProxyConnection connection = PoolConnection.INSTANCE.getConnection()) {
-			statement = connection.prepareStatement(GET_REQUEST_STATUS_LIST);
+			statement = connection.prepareStatement(GET_ALL_PARAMETER_HEADERS);
 			if (statement != null) {
 				resultSet = statement.executeQuery();
 				while (resultSet.next()) {
-					requestStatus.add(buildRequestStatus(resultSet));
+					parameterHeaders.add(buildParameterHeader(resultSet));
 				}
 			}
 		} catch (SQLException e) {
-			throw new DAOException("Cannot load request status list", e);
+			throw new DAOException("Cannot load parameter header list", e);
 		} finally {
 			closeStatement(statement);
 			closeResultSet(resultSet);
 		}
-		return requestStatus;
+		return parameterHeaders;
 	}
 
 	@Override
-	public RequestStatus getRequestStatusByValue(String value) throws DAOException {
-		RequestStatus requestStatus = null;
+	public ParameterHeader getParameterHeaderByValue(String value) throws DAOException {
+		ParameterHeader parameterHeader = null;
 		PreparedStatement statement = null;
 		ResultSet resultSet = null;
 		try (ProxyConnection connection = PoolConnection.INSTANCE.getConnection()) {
-			statement = connection.prepareStatement(GET_REQUEST_STATUS_BY_VALUE);
+			statement = connection.prepareStatement(GET_PARAMETER_HEADER_BY_NAME);
 			if (statement != null) {
 				statement.setString(1, value);
 				resultSet = statement.executeQuery();
 				if (resultSet.next()) {
-					requestStatus = buildRequestStatus(resultSet);
+					parameterHeader = buildParameterHeader(resultSet);
 				}
 			}
 		} catch (SQLException e) {
-			throw new DAOException("Cannot load request status", e);
+			throw new DAOException("Cannot load parameter header", e);
 		} finally {
 			closeStatement(statement);
 			closeResultSet(resultSet);
 		}
-		return requestStatus;
+		return parameterHeader;
 	}
-
 }
