+++ src/main/java/by/training/payment/dao/impl/PaymentSystemDAOImpl.java
@@ -6,60 +6,59 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-import by.training.payment.dao.UserRoleDAO;
-import by.training.payment.entity.UserRole;
+import by.training.payment.dao.PaymentSystemDAO;
+import by.training.payment.entity.PaymentSystem;
 import by.training.payment.exception.DAOException;
 import by.training.payment.pool.PoolConnection;
 import by.training.payment.pool.ProxyConnection;
 
-public class MySQLUserRoleDAO extends SQLUtil implements UserRoleDAO {
+public class PaymentSystemDAOImpl extends SQLUtil implements PaymentSystemDAO {
 
-	private static final String GET_ALL_USER_ROLES = "SELECT * FROM user_role";
-	private static final String GET_USER_ROLE_BY_VALUE = "SELECT * FROM user_role WHERE user_role_value = ?";
+	private static final String GET_PAYMENT_SYSTEM_BY_NAME = "SELECT * FROM payment_system WHERE payment_system_name = ?";
+	private static final String GET_ALL_PAYMENT_SYSTEMS = "SELECT * FROM payment_system";
 
 	@Override
-	public List<UserRole> getAllUserRoles() throws DAOException {
-		List<UserRole> userRoleList = new ArrayList<>();
+	public List<PaymentSystem> getAllPaymentSystems() throws DAOException {
+		List<PaymentSystem> paymentSystems = new ArrayList<>();
 		PreparedStatement statement = null;
 		ResultSet resultSet = null;
 		try (ProxyConnection connection = PoolConnection.INSTANCE.getConnection()) {
-			statement = connection.prepareStatement(GET_ALL_USER_ROLES);
+			statement = connection.prepareStatement(GET_ALL_PAYMENT_SYSTEMS);
 			if (statement != null) {
 				resultSet = statement.executeQuery();
 				while (resultSet.next()) {
-					userRoleList.add(buildUserRole(resultSet));
+					paymentSystems.add(buildPaymentSystem(resultSet));
 				}
 			}
 		} catch (SQLException e) {
-			throw new DAOException("Cannot load user role list", e);
+			throw new DAOException("Cannot load payment system list", e);
 		} finally {
 			closeStatement(statement);
 			closeResultSet(resultSet);
 		}
-		return userRoleList;
+		return paymentSystems;
 	}
 
 	@Override
-	public UserRole getUserRoleByValue(String value) throws DAOException {
-		UserRole userRole = null;
+	public PaymentSystem getPaymentSystemByName(String name) throws DAOException {
+		PaymentSystem paymentSystem = null;
 		PreparedStatement statement = null;
 		ResultSet resultSet = null;
 		try (ProxyConnection connection = PoolConnection.INSTANCE.getConnection()) {
-			statement = connection.prepareStatement(GET_USER_ROLE_BY_VALUE);
+			statement = connection.prepareStatement(GET_PAYMENT_SYSTEM_BY_NAME);
 			if (statement != null) {
-				statement.setString(1, value);
+				statement.setString(1, name);
 				resultSet = statement.executeQuery();
 				if (resultSet.next()) {
-					userRole = buildUserRole(resultSet);
+					paymentSystem = buildPaymentSystem(resultSet);
 				}
 			}
 		} catch (SQLException e) {
-			throw new DAOException("Cannot load user role by value", e);
+			throw new DAOException("Cannot load payment system", e);
 		} finally {
 			closeStatement(statement);
 			closeResultSet(resultSet);
 		}
-		return userRole;
+		return paymentSystem;
 	}
-
 }
